PollSelectorProvider: Tweak existing change
Instead of inheriting the method inheritedChannel() from
SelectorProvider, duplicate the super implementation by
explicitly returning null; this seems like a more robust
replacement for the upstream call to InheritedChannel.getChannel()
(which, if it was present on Android, would always return null
for the reason given in the change summary).

This is not a behavior change, since the super implementation
was already returning null.

Test: Treehugger
Bug: 74377400

Change-Id: I3372bb69366acc87af60a741728a63c256161dfb
Signed-off-by: mydongistiny <jaysonedson@gmail.com>