Update MethodHandleInfo/Natives to OpenJDK 8u121-b13
Verify j.l.invoke.MethodHandleInfo/Natives against OpenJDK 8u121-b13

The MethodHandleInfo class has a number of patches to inline constants
and methods from the upstream MethodHandleNatives class instead of just
including the MethodHandleNatives class. This change adds the
MethodHandleNatives class and reverts those patches. The
MethodHandleNatives class does require some patches in order to remove
code that does not compile or which is otherwise unused. However,
patches that remove code are generally less risky and easier to
maintain than patches that change the code so this seems like a better
trade off.

The refKind...() methods in MethodHandleInfo could not be removed
because they were accidentally added to the public API in API level 26.
The removal of those methods can be tracked under bug 114175595.

Verify java.lang.invoke.MethodHandle/Impl against OpenJDK 8u121-b13

Changes were non-functional and simply added/updated comments to
provide more details of the changes that were made.

The MethodHandleImpl is not based on the upstream code at all and so
has its own copyright message and a single change marker at the top of
the class explaining it.

Bug: 74379469
Test: make checkbuild, flash and start up device
Change-Id: I7ce80e178ab47d81c893fbb684bf5a0d82625896
Signed-off-by: mydongistiny <jaysonedson@gmail.com>